fixes #257
