+++ app/src/main/java/org/openobservatory/ooniprobe/common/OrchestraTask.java
@@ -18,9 +18,6 @@ import io.ooni.mk.MKOrchestraTask;
 import io.ooni.mk.MKResourcesManager;
 
 public class OrchestraTask extends AsyncTask<Void, Void, Void> {
-	public static final String WIFI = "wifi";
-	public static final String MOBILE = "mobile";
-	public static final String NO_INTERNET = "no_internet";
 	private Application app;
 
 	public OrchestraTask(Application app) {
@@ -47,7 +44,7 @@ public class OrchestraTask extends AsyncTask<Void, Void, Void> {
 			client.setGeoIPCountryPath(MKResourcesManager.getCountryDBPath(app));
 			client.setGeoIPASNPath(MKResourcesManager.getASNDBPath(app));
 			client.setLanguage(Locale.getDefault().getLanguage());
-			client.setNetworkType(getNetworkType(app));
+			client.setNetworkType(ReachabilityManager.getNetworkType(app));
 			client.setPlatform("android");
 			//TODO ORCHESTRATE - TIMEZONE
 			//client.setProbeTimezone(TimeZone.getDefault().getDisplayName(true, TimeZone.SHORT));
@@ -67,20 +64,6 @@ public class OrchestraTask extends AsyncTask<Void, Void, Void> {
 		return supportedTest;
 	}
 
-	public static String getNetworkType(Context context) {
-		ConnectivityManager connectivityManager = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
-		if (connectivityManager == null)
-			return NO_INTERNET;
-		else {
-			NetworkInfo info = connectivityManager.getActiveNetworkInfo();
-			if (info != null && info.getType() == ConnectivityManager.TYPE_WIFI)
-				return WIFI;
-			else if (info != null && info.getType() == ConnectivityManager.TYPE_MOBILE)
-				return MOBILE;
-			else
-				return NO_INTERNET;
-		}
-	}
 
 	@Override protected Void doInBackground(Void... voids) {
 		sync(app);
